arcu: simple arcanist wrapper to deal with stacked diffs
Gerrit/git-review make it extremely easy to work with patchset that depend on each
other. arcanist technically supports this workflow but it requires a little more
handholding which quickly becomes tedious for the end-user (who bothers remembering
diff ids?)

arcu is a small wrapper around 'arc diff' that automatically figures out
the phabricator diff id of the current commit as well as the base commit
to update.